server/requirements.txt
Django>=5.2
gunicorn
django-prometheus
djangorestframework
pyyaml
requests
django-cors-headers
celery>=5.0.0
prometheus-client
